requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1

# Templates & Frontend
jinja2==3.1.3
python-multipart==0.0.9

# Excel Processing
openpyxl==3.1.2

# Image Generation
html2image==2.0.4.3
Pillow==10.2.0

# Google Drive API
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Configuration
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Utils
python-dateutil==2.8.2
uuid==1.30
